+++ packages/leancode_analytics/lib/core/core.dart
@@ -1,3 +1,0 @@
-export 'analytics.dart';
-export 'event.dart';
-export 'route_observer.dart';