+++ packages/leancode_analytics/lib/firebase/firebase.dart
@@ -1,1 +1,0 @@
-export 'firebase_analytics.dart';