+++ packages/leancode_analytics_base/lib/src/analytics.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/widgets.dart';
+
+import 'event.dart';
+
+abstract interface class LeanAnalytics {
+  void register(AnalyticsEvent event);
+
+  List<NavigatorObserver> get navigatorObservers;
+}