+++ packages/leancode_analytics_base/lib/src/route.dart
@@ -1,0 +1,5 @@
+abstract interface class LeanAnalyticsRoute {
+  String get id;
+
+  Map<String, dynamic>? get analyticsParams;
+}