+++ packages/leancode_analytics_posthog/lib/leancode_analytics_posthog.dart
@@ -1,0 +1,4 @@
+export 'package:leancode_analytics_base/leancode_analytics_base.dart';
+export 'package:posthog_flutter/posthog_flutter.dart';
+
+export 'src/posthog_analytics.dart';